error v1

+++ app/(routes)/success/page.tsx
@@ -3,7 +3,9 @@ import { clearCart } from "@/app/GlobalRedux/Features/counter/counterSlice";
 import { useDispatch } from "react-redux";
 
 const Page = () => {
-  localStorage.clear();
+  if (localStorage.getItem("cart")) {
+    localStorage.clear();
+  }
   const dispatch = useDispatch();
   dispatch(clearCart());
   return (
